requirements.txt
dj-database-url==0.5.0
Django==2.1.8
django-bootstrap-modal-forms==1.4.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-heroku==0.3.1
django-widget-tweaks==1.4.2
djangorestframework==3.10.3
gunicorn==19.9.0
html5lib==1.0.1
Pillow==6.1.0
psycopg2==2.7.7
psycopg2-binary==2.8.2
PyPDF2==1.26.0
pytz==2018.5
reportlab==3.5.26
selenium==3.14.0
six==1.12.0
urllib3==1.25.3
webencodings==0.5.1
whitenoise==4.1.4
xhtml2pdf==0.2.3
